Moved custom navigation to separate class
